Add oval patterns, refactor css
I moved a ruleblock from region.css to curve.css for better separation
of concerns.

Previously, the rule only applies styling to dividers that are inside a
`.region` element. I removed that restriction, and because the rule is
no longer attached to the region utility class, it makes sense to move
its location.